thread set to 1, update all to int, et logging logic updated

+++ apps/ges/inc/ring_buffer.h
@@ -8,15 +8,15 @@
 typedef struct
 {
 	// double buffer[SIGNAL_PROCESSING_BUFFER_SIZE]; // Fixed-size buffer
-	int16_t *buffer;				 // Pointer to the buffer array
-	int16_t size;						 // Size of the buffer array
-	int16_t *head;					 // Write pointer
-	int16_t *tail;					 // Read pointer
-	int16_t *end;						 // Pointer to end of the buffer(next address after the last element))
-	int8_t is_full;					 // Flag to indicate if the buffer is full
-	int8_t rtr_flag;				 // Flag to indicate if the buffer is ready
-	int8_t new_signal_count; // count of write after last read
-	int32_t cur_time_ms;		 // Current time in milliseconds (for debugging or logging purposes)
+	int *buffer;					// Pointer to the buffer array
+	int size;							// Size of the buffer array
+	int *head;						// Write pointer
+	int *tail;						// Read pointer
+	int *end;							// Pointer to end of the buffer(next address after the last element))
+	int is_full;					// Flag to indicate if the buffer is full
+	int rtr_flag;					// Flag to indicate if the buffer is ready
+	int new_signal_count; // count of write after last read
+	int cur_time_ms;			// Current time in milliseconds (for debugging or logging purposes)
 } RingBuffer;
 
 /**
@@ -26,7 +26,7 @@ typedef struct
  * @param buffer Pointer to the buffer array
  * @param buffer_size Size of the buffer array
  */
-void rb_init(RingBuffer *rb, int16_t *buffer, int16_t buffer_size);
+void rb_init(RingBuffer *rb, int *buffer, int buffer_size);
 
 /**
  * Pushes a new sample into the ring buffer
@@ -36,7 +36,7 @@ void rb_init(RingBuffer *rb, int16_t *buffer, int16_t buffer_size);
  * @param cur_time_ms Pointer to the current time in milliseconds (to be updated)
  * @return 1 on success, 0 on failure
  */
-int rb_push_sample(RingBuffer *rb, int16_t new_sample);
+int rb_push_sample(RingBuffer *rb, int new_sample);
 
 /**
  * Takes a snapshot of the ring buffer
@@ -46,7 +46,7 @@ int rb_push_sample(RingBuffer *rb, int16_t new_sample);
  * @param offset Offset in the ring buffer from which to start taking the snapshot
  * @return 1 on success, 0 on failure
  */
-int rb_snapshot(RingBuffer *rb, int16_t *snapshot_buffer, int16_t offset);
+int rb_snapshot(RingBuffer *rb, int *snapshot_buffer, int offset);
 
 /**
  * Resets the ring buffer
